requirements.txt
mediapipe
opencv-python
pandas
numpy
scipy
jupyter
pympi-ling